DiscordBot/requirements.txt
hikari >= 2.5.0
hikari-lightbulb >= 3.2.3